requirements.txt
numpy
tk-tools
pyueye
pandas
tqdm
argparse
matplotlib
Pillow
pyueye
PyQt5
opencv-python=4.8.0.74
opencv-contrib-python=4.8.1.78
scikit-learn=1.3.2
tensorflow=2.11.0